+++ AoC2019/Common/IntCodeComputer/Instructions/Output.cs
@@ -4,10 +4,10 @@ namespace AoC2019.Common.IntCodeComputer.Instructions
 {
     public struct Output : IInstruction
     {
-        private readonly Action<int> _setOutput;
+        private readonly Action<long> _setOutput;
         public int Length => 2;
 
-        public Output(Action<int> setOutput)
+        public Output(Action<long> setOutput)
         {
             _setOutput = setOutput;
         }
